+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 5.1.0
 -----
 
+ * deprecated support for `service:action` syntax to reference controllers, use `serviceOrFqcn::method` instead
  * allowed using public aliases to reference controllers
  * added session usage reporting when the `_stateless` attribute of the request is set to `true`
  * added `AbstractSessionListener::onSessionUsage()` to report when the session is used while a request is stateless
@@ -12,8 +13,8 @@ CHANGELOG
 -----
 
  * removed support for getting the container from a non-booted kernel
- * removed the first and second constructor argument of `ConfigDataCollector` 
- * removed `ConfigDataCollector::getApplicationName()` 
+ * removed the first and second constructor argument of `ConfigDataCollector`
+ * removed `ConfigDataCollector::getApplicationName()`
  * removed `ConfigDataCollector::getApplicationVersion()`
  * removed support for `Symfony\Component\Templating\EngineInterface` in `HIncludeFragmentRenderer`, use a `Twig\Environment` only
  * removed `TranslatorListener` in favor of `LocaleAwareListener`
@@ -25,7 +26,7 @@ CHANGELOG
  * removed `GetResponseForControllerResultEvent`, use `ViewEvent` instead
  * removed `GetResponseForExceptionEvent`, use `ExceptionEvent` instead
  * removed `PostResponseEvent`, use `TerminateEvent` instead
- * removed `SaveSessionListener` in favor of `AbstractSessionListener` 
+ * removed `SaveSessionListener` in favor of `AbstractSessionListener`
  * removed `Client`, use `HttpKernelBrowser` instead
  * added method `getProjectDir()` to `KernelInterface`
  * removed methods `serialize` and `unserialize` from `DataCollector`, store the serialized state in the data property instead