+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 5.1.0
 -----
 
+ * Allow request locale and Response `Content-Language` header to be set from `preferredLocale` thanks to `available_locales`
  * made `WarmableInterface::warmUp()` return a list of classes or files to preload on PHP 7.4+;
    not returning an array is deprecated
  * deprecated support for `service:action` syntax to reference controllers, use `serviceOrFqcn::method` instead